development: Updated README Features menu
Added "Feature #" column to Features menu in README.  This is to facilitate name branching when it comes time to start implementing features.
I'm kind of early with this but I feel like it won't take much to implement the ZfsFile class since I'm not writing unit tests for it.

Changes to be committed:
	modified:   README.md