development: Functional Test update
Implemented Backup functionality (since we're editing files now)
Implemented a section to remake the original save game.
All tests currently passing but the *real* functional test is causing the in-game load to crash.  I suspect it's due to the method of compression I'm using.  Time for a refactor to store the original compress_type from ZipInfo and then put it back the way we found it.

Changes to be committed:
	modified:   zks_file_class_test.py