development: Case Study files for load bug
I was using these files to try and troubleshoot the difference between what the game wants and what the game gets.
The original is just that, it's the player.json from a known-good, unaltered save.
The rest of the broken?.json files come from ZksFile functional tests.

Changes to be committed:
	new file:   Test_Files/Case_Study/player(broken1).json
	new file:   Test_Files/Case_Study/player(broken2).json
	new file:   Test_Files/Case_Study/player(broken3).json
	new file:   Test_Files/Case_Study/player(broken4).json
	new file:   Test_Files/Case_Study/player(broken5).json
	new file:   Test_Files/Case_Study/player(broken6).json
	new file:   Test_Files/Case_Study/player(original).json