development: Init. Commit JsonFile class raw tests
Semi-dirty test file to do some heavy lifing to compare known-good player.json files and my edited player.json files.  Coupled with xxd dumps, I was able to use this script to work out some bugs.  Most importantly, I was able to discover that the player.json file expects a certain order.  Finally, I used it as PoC code to verify that an ordered dict would correctly interface with the json module to maintain the order of a json.loads/dumps.

Changes to be committed:
	new file:   json_file_class_test_raw_compare.py