development: Implemented dynamic user menu
Implemented a dynamic user menu in baron_builder.py.  user_mod_menu() calls the *_availabile() functions from the 'features' module to determine what is available in a given save game (see: save game maturity).  Available features are loaded into a container and the user menu is dynamically created from that.  It appears to be entirely functional.

Changes to be committed:
	modified:   baron_builder.py