F05: Updated README based on work done
GOOD NEWS: Updated Feature 05 based on Linux and Windows functional testing.
BAD NEWS: No one should be using it yet because there's no archive bookkeeping yet (see: F04)

PLOT TWIST: The "Load Games" menu in PKM displays whatever is in the Save Game directory, regardless of whether or not steam-saves-release.json has an entry for it or not.  WHY?!  Why keep a running list of all the save games and their versions only to display whatever is in the folder?!
FUTURE: I added a Feature to the table to clean up steam-saves-release.json based on what files it finds in the save game directory.  Got a steam-saves-release.json with no corresponding save game?  Remove the steam-saves-release.json entry!

NOTE: Merge request imminent.  SHIP IT!

Changes to be committed:
	modified:   README.md