F04: Final(?) b_b_file_mgmt.py update
The big changes are as follows:

- Implemented check_filename_no_ext() as a helper function
- Fixed BUG in add_save_game_to_list()  (Previously, it didn't account for entries that already existed)
- backup_a_file()
	- Fixed BUG(?) (Added better return value validation)
	- Extricated start_storage_dir() up one level to user_file_menu() so that I could use this function to UNbackup_a_file()
- user_file_menu()
	- Implemented option D for backup saves
	- Added backup file list cleanup to the end

GOOD NEWS: This feature has been throughly tested on Linux:
	- Restored over existing file
	- Restored missing file, json entry present
	- Restored over file, json entry missing
	- Restored verifying correction of json entry
BAD NEWS: I haven't been able to test this on a Windows machine.  I'll document it as such but I'm submitting a merge request regardless.

NOTE: Modifying this option to also include archive saves (see: Feature 12) may be tricky but it's possible.

Changes to be committed:
	modified:   baron_builder_file_mgmt.py